+++ tests/testthat/test-hash.R
@@ -50,10 +50,8 @@ test_with_dir("same with a directory", {
 
 test_with_dir("hashing decisions", {
   skip_on_cran() # CRAN gets whitelist tests only (check time limits).
-  file <- "nobodyhome"
   expect_true(
     should_rehash_storage(
-      filename = file,
       new_mtime = 0,
       old_mtime = 0,
       old_size = 0,
@@ -64,7 +62,6 @@ test_with_dir("hashing decisions", {
   for (i in c(0, 1)) {
     expect_false(
       should_rehash_storage(
-        filename = file,
         new_mtime = 0,
         old_mtime = i,
         old_size = 0,
@@ -74,7 +71,6 @@ test_with_dir("hashing decisions", {
     )
     expect_true(
       should_rehash_storage(
-        filename = file,
         new_mtime = 0,
         old_mtime = i,
         old_size = 0,
@@ -86,7 +82,6 @@ test_with_dir("hashing decisions", {
   for (s in c(-Inf, Inf)) {
     expect_true(
       should_rehash_storage(
-        filename = file,
         new_mtime = 1,
         old_mtime = 0,
         old_size = 0,
@@ -96,7 +91,6 @@ test_with_dir("hashing decisions", {
     )
     expect_true(
       should_rehash_storage(
-        filename = file,
         new_mtime = 0,
         old_mtime = 0,
         old_size = 1,
@@ -106,7 +100,6 @@ test_with_dir("hashing decisions", {
     )
     expect_true(
       should_rehash_storage(
-        filename = file,
         new_mtime = 0,
         old_mtime = 0,
         old_size = 0,